+++ RELEASE_NOTES.md
@@ -6,9 +6,12 @@ In development.
 
 - Added DataNode.copy() method.
 - Added Node.findDescendant() method.
+- Added default methods Parser.parse(File) and Parser.parse(String).
+- Added convenience methods SDA.parse(File) and SDA.parse(String).
 - Added default methods Formatter.format(File, Node) and Formatter.format(Node).
 - Added convenience methods SDA.format(File, DataNode) and SDA.format(DataNode).
 - Closed issue #16 (SDAFormatter throws if input node is null).
+- Closed issue #17 (Add a parse(File) method and review exceptions).
 
 ## Compatibility